Added the entire coding process for creating and evaluating the KNN classification model.
I split the data, created a recipe, tuned the spec model, selected the best K value, and then evaluated the best spec model with the testing data. The accuracy is quite low, reflecting the lack of class boundaries in the data. I will add writing that accompanies the coding steps later.